[TASK] Initial commit of code base
This contains basic functionality, however serving only
as suggestions.

Shipped with this is a Facebook flow implementation, some
additional commonly used ones should be added, like G+.

Still to check is how the actual handling on authentication
should be, especially if there already is a party, considering
the same user, but with different authentications.